requirements.txt
Flask==0.10.1

# Flask Extensions
Flask-Assets==0.10
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.0
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.11
Flask-User==0.6.8
Flask-Bootstrap==3.3.6.0

# Other
itsdangerous==0.24
cssmin==0.2.0
jsmin==2.1.1
WTForms-Alchemy==0.16.1

# Testing
pytest==2.7.0
pytest-cov==1.8.1
mccabe==0.3
flake8==2.4.0
